Rename private classes
