+++ azure-kusto-ingest/azure/kusto/ingest/__init__.py
@@ -1,6 +1,6 @@
 from .kusto_ingest_client import KustoIngestClient
-from .descriptors import BlobDescriptor, FileDescriptor
-from .ingestion_properties import (
+from ._descriptors import BlobDescriptor, FileDescriptor
+from ._ingestion_properties import (
     DataFormat,
     ValidationPolicy,
     ValidationImplications,
