+++ azure-kusto-ingest/azure/kusto/ingest/_connection_string.py
