+++ azure-kusto-ingest/azure/kusto/ingest/_resource_manager.py
@@ -2,7 +2,7 @@
 
 from datetime import datetime, timedelta
 
-from .connection_string import _ConnectionString
+from ._connection_string import _ConnectionString
 
 
 class _IngestClientResources:
